[FIX] Fixed status and navigation bar theming issue.
