+++ app/src/main/java/chat/rocket/android/main/presentation/MainNavigator.kt
@@ -11,8 +11,7 @@ import chat.rocket.android.main.ui.MainActivity
 import chat.rocket.android.profile.ui.TAG_PROFILE_FRAGMENT
 import chat.rocket.android.server.ui.changeServerIntent
 import chat.rocket.android.settings.ui.TAG_SETTINGS_FRAGMENT
-import chat.rocket.android.thememanager.ui.TAG_THEME_FRAGMENT
-import chat.rocket.android.thememanager.ui.newInstance
+import chat.rocket.android.thememanager.ui.ThemesActivity
 import chat.rocket.android.util.extensions.addFragment
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import chat.rocket.android.webview.adminpanel.ui.TAG_ADMIN_PANEL_WEB_VIEW_FRAGMENT
@@ -50,13 +49,11 @@ class MainNavigator(internal val activity: MainActivity) {
             chat.rocket.android.profile.ui.newInstance()
         }
     }
-    //----------------------------------------------------------------------------------
+
     fun toChangeTheme() {
-        activity.addFragmentBackStack(TAG_THEME_FRAGMENT, R.id.fragment_container) {
-            newInstance()
-        }
+        activity.startActivity(Intent(activity, ThemesActivity::class.java))
+        activity.finish()
     }
-    //---------------------------------------------------------------------------------
 
     fun toAdminPanel(webPageUrl: String, userToken: String) {
         activity.addFragmentBackStack(TAG_ADMIN_PANEL_WEB_VIEW_FRAGMENT, R.id.fragment_container) {
